// src/app/models/studentRegistration.model.ts
export class studentRegistration{
    StudentName?: string ;
    LastRegNumber?: string ;
    SelectedClass?: string ;
    AdmissionDate?: string ;
    Discount?: number ;
    MobileNumber?: string ;
    Dob?: string ;
    StudentNationalID?: string ;
    IsOrphan?: string ;
    Gender?: string ;
    Cast?: string ;
    Osc?: string ;
    Identity?: string ;
    PreviousSchool?: string ;
    Religion?: string ;
    BloodGroup?: string ;
    PreviousRollNumber?: string ;
    Family?: string ;
    Disease?: string ;
    Notes?: string ;
    Siblings?: string ;
    Address?: string ;
    FatherName?: string ;
    FatherNationalID?: string ;
    FatherOccupation?: string ;
    FatherEducation?: string ;
    FatherMobileNumber?: string ;
    FatherProfession?: string ;
    FatherIncome?: string ;
    MotherName?: string ;
    MotherNationalID?: string ;
    MotherOccupation?: string ;
    MotherEducation?: string ;
    MotherMobileNumber?: string ;
    MotherProfession?: string ;
    MotherIncome?: string ;
    id?: string;
} 